Rename reused components to base for clarity.
